remove unecessary blocking call
After publishing a message to pubsub we were waiting for the
`result.Ready()` channel to be closed. However, this is already properly
handle by the following call to `result.Get()` function.

This removes the unecessary blocking operation waiting for the
`result.Ready()` channel, once this is already handled by a call to the
`result.Get()` function.