propagate tracing context to the pubsub sdk
The google pubsub SDK expects tracing information to be present in
provided context when open telemetry tracing is enabled.

Even if the `message.Message` contains valid tracing information in
`Metadata` and we properly convert it into pubsub message `Attributes`,
that is not taken into account when propagating the span, so that the
traceparent Attribute gets overwritten with a new root span.

The result of this is that we get totally disconnected span, once the
trace id is not propgated.

This extracts tracing information from `message.Message.Metadata` and
injects it into the context before passing it to the `Publish` method of
the SDK, allowing for proper tracing between distributed components.